agents: Create the ``lsb`` agent
The ``lsb`` agent provides with `Linux Standard Base (LSB)`_ compliance
in order to find back common utilities between the Linux distributions.

It is not the objective to have the whole standard installed but only a
subset of it. Currently the agent installs the ``lsb_release`` utility
and reports about the distribution.

.. _Linux Standard Base (LSB): http://www.linuxfoundation.org/collaborate/workgroups/lsb

Signed-off-by: Jimmy Thrasibule <jimmy.thrasibule@gmail.com>